backend/requirements.txt
asgiref
Django
djangorestframework
python-dotenv
psycopg2-binary
gunicorn
djoser
drf-extra-fields
Pillow
dotenv
django-filter